Move two HTML building fns to _utils_html.py
